requirements.txt
Flask==2.3.2
pandas==1.5.3
plotly==5.9.0
numpy==1.23.1
openai==0.9.0
python-dotenv==0.21.0
Flask-Cors==3.1.0
